src: ocl: Improve INT8 performance on GEN
src: ocl: Apply clang format

+++ src/ocl/gen12lp_gemm_nocopy_scale_x8x8s32.cl
@@ -37,17 +37,17 @@
     } while (0)
 
 kernel void gen12lp_gemm_scale_x8x8s32_kernel(global int *cc, global int *c,
-        char trc, long offset_c, long m, long n, long ldc, float alpha,
-        float beta, global int *co, long offset_co, int alpha_is_zero,
-        int apply_eltwise, float eltwise_alpha, float eltwise_beta) {
+        char trc, int offset_c, int m, int n, int ldc, float alpha, float beta,
+        global int *co, int offset_co, int alpha_is_zero, int apply_eltwise,
+        float eltwise_alpha, float eltwise_beta) {
 
     int idx = get_group_id(0);
     int idy = get_group_id(1);
     int lid = get_local_id(0);
-    long j;
-    long offset_cc = 0;
-    long offset_x = 0;
-    long ldcc = m;
+    int j;
+    int offset_cc = 0;
+    int offset_x = 0;
+    int ldcc = m;
 
     m -= 32 * idx;
     if (m > 32) m = 32;
@@ -63,11 +63,11 @@ kernel void gen12lp_gemm_scale_x8x8s32_kernel(global int *cc, global int *c,
         if (m > 0) {
             if (!alpha_is_zero) {
                 c[offset_c + 0]
-                        = (int)((double)alpha * cc[offset_cc + 0]
-                                  + (double)beta * c[offset_c + 0] + (0.0))
+                        = (int)((float)alpha * cc[offset_cc + 0]
+                                  + (float)beta * c[offset_c + 0] + (0.0))
                         + co[offset_co + offset_x];
             } else {
-                c[offset_c + 0] = (int)((double)beta * c[offset_c + 0] + (0.0))
+                c[offset_c + 0] = (int)((float)beta * c[offset_c + 0] + (0.0))
                         + co[offset_co + offset_x];
             }
             if (trc == 'C') { offset_x++; }
@@ -76,11 +76,11 @@ kernel void gen12lp_gemm_scale_x8x8s32_kernel(global int *cc, global int *c,
         if (m > 1) {
             if (!alpha_is_zero) {
                 c[offset_c + 1]
-                        = (int)((double)alpha * cc[offset_cc + 1]
-                                  + (double)beta * c[offset_c + 1] + (0.0))
+                        = (int)((float)alpha * cc[offset_cc + 1]
+                                  + (float)beta * c[offset_c + 1] + (0.0))
                         + co[offset_co + offset_x];
             } else {
-                c[offset_c + 1] = (int)((double)beta * c[offset_c + 1] + (0.0))
+                c[offset_c + 1] = (int)((float)beta * c[offset_c + 1] + (0.0))
                         + co[offset_co + offset_x];
             }
             if (apply_eltwise) UPDATE_ELEM_C(offset_c + 1);
